:art: Reorganise the commands
