+++ src/rogallo/commands/main.py
@@ -6,48 +6,6 @@ from textual_enhanced.commands import Command
 
 
 ##############################################################################
-class JumpToCommandLine(Command):
-    """Jump to the command line"""
-
-    BINDING_KEY = "/, ctrl+1"
-
-
-##############################################################################
-class JumpToDocument(Command):
-    """Jump to the document viewer"""
-
-    BINDING_KEY = "ctrl+slash, ctrl+g, ctrl+2"
-
-
-##############################################################################
-class JumpToSidebar(Command):
-    """Jump to the sidebar"""
-
-    BINDING_KEY = "ctrl+3"
-
-
-##############################################################################
-class ChangeCommandLineLocation(Command):
-    """Swap the position of the command line between top and bottom"""
-
-    BINDING_KEY = "ctrl+up, ctrl+down"
-
-
-##############################################################################
-class ToggleHistoryManager(Command):
-    """Toggle the display of the history viewer"""
-
-    BINDING_KEY = "shift+f2"
-
-
-##############################################################################
-class ToggleBookmarksManager(Command):
-    """Toggle the display of the bookmarks viewer"""
-
-    BINDING_KEY = "shift+f3"
-
-
-##############################################################################
 class Reload(Command):
     """Reload the current document"""
 
@@ -59,13 +17,6 @@ class ToggleView(Command):
     """Toggle between rendered and source view of the document"""
 
     BINDING_KEY = "f4"
-
-
-##############################################################################
-class GoHome(Command):
-    """Go to the home page"""
-
-    BINDING_KEY = "ctrl+h"
 
 
 ##############################################################################
@@ -94,42 +45,6 @@ class ClearCache(Command):
     """Clear the cache for all content"""
 
     BINDING_KEY = "shift+f5"
-
-
-##############################################################################
-class StripeLinks(Command):
-    """Toggle the striping of links in the document viewer"""
-
-    BINDING_KEY = "f8"
-
-
-##############################################################################
-class ToggleLinkNumbers(Command):
-    """Toggle the display of link numbers in the document viewer"""
-
-    BINDING_KEY = "shift+f8"
-
-
-##############################################################################
-class ToggleCosyLinkNumbers(Command):
-    """Toggle the position of link numbers when they're being displayed"""
-
-    BINDING_KEY = "super+f8"
-
-
-##############################################################################
-class ToggleEmojiRemoval(Command):
-    """Toggle the removal of emoji from text content"""
-
-    BINDING_KEY = "f6"
-
-
-##############################################################################
-class ToggleANSIEscapeSequenceHandling(Command):
-    """Toggle the handling of ANSI escape sequences in text content"""
-
-    BINDING_KEY = "shift+f6"
-    ACTION = "toggle_ansi_escape_sequence_handling_command"
 
 
 ##############################################################################
